Batch request in json format -- WIP
