+++ src/Microsoft.OData.Core/Microsoft.OData.Core.Portable45.Profile111.csproj
@@ -255,6 +255,7 @@
     <Compile Include="ODataBatchOperationWriteStream.cs" />
     <Compile Include="ODataBatchReader.cs" />
     <Compile Include="ODataBatchJsonReader.cs" />
+    <Compile Include="ODataBatchMimeReader.cs" />
     <Compile Include="ODataBatchReaderState.cs" />
     <Compile Include="ODataBatchReaderStream.cs" />
     <Compile Include="ODataBatchReaderStreamBuffer.cs" />
