+++ src/Microsoft.OData.Core/ODataRawInputContext.cs
@@ -209,7 +209,7 @@ namespace Microsoft.OData.Core
         /// <returns>The newly created <see cref="ODataCollectionReader"/>.</returns>
         private ODataBatchReader CreateBatchReaderImplementation(string batchBoundary, bool synchronous)
         {
-            return new ODataBatchReader(this, batchBoundary, this.encoding, synchronous);
+            return new ODataBatchMimeReader(this, batchBoundary, this.encoding, synchronous);
         }
 
         /// <summary>
