+++ test/FunctionalTests/Tests/DataOData/Tests/OData.Common.Tests/PublicApi/PublicApi.bsl
@@ -3895,6 +3895,23 @@ public abstract class Microsoft.OData.Core.ODataAnnotatable {
 	public void SetAnnotation (T annotation)
 }
 
+public abstract class Microsoft.OData.Core.ODataBatchReader : IODataBatchOperationListener {
+	Microsoft.OData.Core.ODataBatchReaderState State  { public get; }
+
+	public Microsoft.OData.Core.ODataBatchOperationRequestMessage CreateOperationRequestMessage ()
+	public System.Threading.Tasks.Task`1[[Microsoft.OData.Core.ODataBatchOperationRequestMessage]] CreateOperationRequestMessageAsync ()
+	internal abstract Microsoft.OData.Core.ODataBatchOperationRequestMessage CreateOperationRequestMessageImplementation ()
+	public Microsoft.OData.Core.ODataBatchOperationResponseMessage CreateOperationResponseMessage ()
+	public System.Threading.Tasks.Task`1[[Microsoft.OData.Core.ODataBatchOperationResponseMessage]] CreateOperationResponseMessageAsync ()
+	internal abstract Microsoft.OData.Core.ODataBatchOperationResponseMessage CreateOperationResponseMessageImplementation ()
+	void Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamDisposed ()
+	void Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamRequested ()
+	System.Threading.Tasks.Task Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamRequestedAsync ()
+	public bool Read ()
+	public System.Threading.Tasks.Task`1[[System.Boolean]] ReadAsync ()
+	internal abstract bool ReadImplementation ()
+}
+
 public abstract class Microsoft.OData.Core.ODataBatchWriter : IODataBatchOperationListener, IODataOutputInStreamErrorListener {
 	public abstract void BatchOperationContentStreamDisposed ()
 	public abstract void BatchOperationContentStreamRequested ()
@@ -4333,20 +4350,6 @@ public sealed class Microsoft.OData.Core.ODataUtils {
 	public static Microsoft.OData.Core.ODataVersion StringToODataVersion (string version)
 }
 
-public class Microsoft.OData.Core.ODataBatchReader : IODataBatchOperationListener {
-	Microsoft.OData.Core.ODataBatchReaderState State  { public get; }
-
-	public Microsoft.OData.Core.ODataBatchOperationRequestMessage CreateOperationRequestMessage ()
-	public System.Threading.Tasks.Task`1[[Microsoft.OData.Core.ODataBatchOperationRequestMessage]] CreateOperationRequestMessageAsync ()
-	public Microsoft.OData.Core.ODataBatchOperationResponseMessage CreateOperationResponseMessage ()
-	public System.Threading.Tasks.Task`1[[Microsoft.OData.Core.ODataBatchOperationResponseMessage]] CreateOperationResponseMessageAsync ()
-	void Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamDisposed ()
-	void Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamRequested ()
-	System.Threading.Tasks.Task Microsoft.OData.Core.IODataBatchOperationListener.BatchOperationContentStreamRequestedAsync ()
-	public bool Read ()
-	public System.Threading.Tasks.Task`1[[System.Boolean]] ReadAsync ()
-}
-
 [
 DebuggerDisplayAttribute(),
 ]
@@ -4429,24 +4432,6 @@ public sealed class Microsoft.OData.Core.ODataAsynchronousWriter : IODataOutputI
 	public System.Threading.Tasks.Task`1[[Microsoft.OData.Core.ODataAsynchronousResponseMessage]] CreateResponseMessageAsync ()
 	public void Flush ()
 	public System.Threading.Tasks.Task FlushAsync ()
-}
-
-public sealed class Microsoft.OData.Core.ODataBatchJsonWriter : Microsoft.OData.Core.ODataBatchWriter, IODataBatchOperationListener, IODataOutputInStreamErrorListener {
-	public virtual void BatchOperationContentStreamDisposed ()
-	public virtual void BatchOperationContentStreamRequested ()
-	public virtual System.Threading.Tasks.Task BatchOperationContentStreamRequestedAsync ()
-	public virtual void Flush ()
-	public virtual System.Threading.Tasks.Task FlushAsync ()
-	public virtual void OnInStreamError ()
-}
-
-public sealed class Microsoft.OData.Core.ODataBatchMimeWriter : Microsoft.OData.Core.ODataBatchWriter, IODataBatchOperationListener, IODataOutputInStreamErrorListener {
-	public virtual void BatchOperationContentStreamDisposed ()
-	public virtual void BatchOperationContentStreamRequested ()
-	public virtual System.Threading.Tasks.Task BatchOperationContentStreamRequestedAsync ()
-	public virtual void Flush ()
-	public virtual System.Threading.Tasks.Task FlushAsync ()
-	public virtual void OnInStreamError ()
 }
 
 public sealed class Microsoft.OData.Core.ODataBatchOperationRequestMessage : IODataRequestMessage, IODataRequestMessageAsync, IODataUrlResolver {
